Fix build break: Python3-ify last change
commit dedb0b6 (adding CAST) landed after 175cec6 (making the code
Python3-compatible) but before the Python 3 testing actually worked
(because of some bazel weirdness), so it unknowingly introduced some
small Python 3 incompatibilities.

One is about an exception being formatted differently, so the test to
check the exception needs to be slightly more general.

The other is more subtle; we were asserting on the pandas series
contained within a TypedSeries returned from a function evaluation,
which left us vulnerable to some changes in string-type handling.  It's
better to test using the public API of the Typed* types, which converts
the result back to Python types from Pandas types.  There actually
wasn't one of these for TypedSeries, only for TypedDataFrame, so,
introduce one and test it.

This handles the strings, but causes new problems for the date/time tests,
which were asserting the result to be a Pandas date/time type
(pd.Timestamp), while the new API returns the corresponding Python type,
datetime or date.  That's probably right, but the asserted return values
need to be changed too.

Change-Id: I419e6457ebad21ddd74ae9c7b45a2da07f84b834