Implement Arrays, and ARRAY and ARRAY_AGG fcns.
Note that Array *types* were already in this implementation, but there
was no syntax and no functions to create array values.

Changed the internal representation of an ARRAY to a tuple rather than a
list; it doesn't need to be mutable, and tuples are what the parser
emits so there's less casting.

Also fix an old bug that made selecting constant expressions in a group
by context not work.  A selector with no group by or aggregation is
invalid in a group by context unless it is constant at compile time, so
we add a check for that case specifically and throw, during group by
(rather than later at evaluation) if not.

Change-Id: I1b71df58cda42179389261b353904ed6cd92658a