Remove redundant function_call code.
Previously the function call grammar rule was a separate function to
deal with the special case of calling a factory function to wrap the
grammar rule result; later this was generalized to a `wrap' function to
wrap any rule, but the specialized function_call rule and function were
not removed.  This change deletes that now-redundant code.

Change-Id: I1a11ebb6bd6a6eb8140c8aaf71f90211bc13e354