Implement STRUCT functions and types
Implement grammar and an ast node allowing the creation of struct
values, and a new BQType subtype representing their types.

Implement a sub-grammar for BQ types, which, with STRUCTs are now
recursive.  Break this out to a separate file in order to avoid
recursive imports.

Smaller changes:
- the Array ast node now takes an array type rather than the member
type, to be able to use the new recursive type parser.  Cast takes a
scalar type rather than a string for symmetry.
- move the wrap helper function to query_helper, so that it can be used
by the type sub-grammar (it probably should have gone there in the first
place).
- add a type signature to the core_expression function, which enables
type-checking there and flags a bunch of things downstream,
necessitating more type assertions in tests that introspect grammar
results.
- add the type grammar to BUILD and run_tests.sh

Not included in this change (will come later since this is already large):
- UNNEST
- accessing struct members in queries

Change-Id: I4f9b109a6904975d8550a6533a7c5262679a7dc9