Move *.py to purplequery/ ; use explicit imports
A statement like 'import foo' intending to import foo.py in the current
directory is a relative import (relative to the current module's
directory) but an implicit one.  Switch to *ex*plicit relative imports
for library files as implicit relative imports are now deprecated, and
absolute imports for test files so that they can be run.

Also add WORKSPACE to make bazel work, and remove some unused junk in
the BUILD file.

Change-Id: I23af456e3595b63b87c0380023d4f4abb45b8820