Fix aggregation by always using groupby
previously an expression like SUM(a+1)+2 wouldn't work outside of a
group by context, because the constants wouldn't properly evaluate to
different-length columns inside and outside of aggregation (the constant
1 inside is a column having the same length as the column `a`, but the
constant 2 outside is a column of length 1, the same as the aggregated
sum).  This is because aggregation in a group by context and out of it
were handled via separate codepaths.  This didn't work and was
unnecessary.  Instead, if any aggregation is present, always use group
by so that aggregation consistently uses a single codepath.

Change-Id: I133acbacf40f56be6eb9d0c3a9f8834dcb646c4d